feat: new dev mode

+++ main.py
@@ -2,7 +2,7 @@ import os
 import shlex
 from model_resolver.cli import main as model_resolver_main
 
-def main(release: str):
+def main(release: str, devmode: bool = True):
     print(f"Release {release}!")
     release = shlex.quote(release)
     if os.path.exists("branch"):
@@ -14,7 +14,7 @@ def main(release: str):
     os.system(f"git clone -b {cloned_branch} {git_url}")
     os.chdir("renders")
     # check if the release tag already exists
-    if os.system(f"git rev-parse --verify --quiet {release}-renders") == 0:
+    if os.system(f"git rev-parse --verify --quiet {release}-renders") == 0 and not devmode:
         print(f"Release {release} already exists!")
         return
 
@@ -27,24 +27,26 @@ def main(release: str):
         output_dir=cwd,
         load_dir=f"{cwd}/resourcepack",
     )
-    os.system("git add .")
-    os.system(f"git commit -m '✨ Generate renders for {release}' --allow-empty")
-    os.system(f"git tag -a '{release}-renders' -m '✨ Generate renders for {release}'")
+    
+    if not devmode:
+        os.system("git add .")
+        os.system(f"git commit -m '✨ Generate renders for {release}' --allow-empty")
+        os.system(f"git tag -a '{release}-renders' -m '✨ Generate renders for {release}'")
 
-    if "GITHUB_TOKEN" in os.environ:
-        token = os.environ["GITHUB_TOKEN"]
-        os.system(f"git remote set-url origin https://github-actions:{token}@{git_url.replace('https://', '')}")
-    os.system("git push origin renders --tags")
-    os.chdir("../..")
-    os.system("rm poetry.lock")
-    os.system("rm -rf branch")
+        if "GITHUB_TOKEN" in os.environ:
+            token = os.environ["GITHUB_TOKEN"]
+            os.system(f"git remote set-url origin https://github-actions:{token}@{git_url.replace('https://', '')}")
+        os.system("git push origin renders --tags")
+        os.chdir("../..")
+        os.system("rm poetry.lock")
+        os.system("rm -rf branch")
 
 
 
 
 
 if __name__ == "__main__":
-    release = os.getenv("MC_VERSION", "24w33a")
+    release = os.getenv("MC_VERSION", "24w45a")
     if release is None or len(release) == 0:
         raise ValueError(f"MC_VERSION is not set, got {release}")
     if not "," in release:
